+++ src/main/java/net/thesquire/backroomsmod/BackroomsMod.java
@@ -29,9 +29,13 @@ import java.util.Objects;
 //TODO redo painted warehouse concrete rebar texture handling
 //TODO add level 1 water puddle feature with dripping water
 //TODO modify GridWalls density function to add passageways
-//TODO add portal feature connecting level 0 to level 1
 //TODO test different smooth noise type for GridWalls density function
 //TODO figure out how to implement regional level 1 dimension blackout events
+//TODO generify portal placer block's destination structure feature
+
+//smaller items
+//TODO add nbt option for portal placer to place portals on edge of block
+//TODO make mountable fluorescent light block break when surface breaks, like torch
 
 public class BackroomsMod implements ModInitializer {
 	public static final String MOD_ID = "backroomsmod";