+++ src/main/java/net/thesquire/backroomsmod/util/ModUtils.java
@@ -1,22 +1,29 @@
 package net.thesquire.backroomsmod.util;
 
+import com.mojang.datafixers.util.Pair;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
+import net.minecraft.block.entity.StructureBlockBlockEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
-import net.minecraft.registry.Registries;
-import net.minecraft.registry.RegistryKey;
-import net.minecraft.registry.RegistryKeys;
+import net.minecraft.registry.*;
 import net.minecraft.registry.entry.RegistryEntry;
+import net.minecraft.registry.entry.RegistryEntryList;
+import net.minecraft.server.world.ServerWorld;
 import net.minecraft.state.State;
 import net.minecraft.state.StateManager;
-import net.minecraft.util.Identifier;
+import net.minecraft.structure.StructurePlacementData;
+import net.minecraft.structure.StructureTemplate;
+import net.minecraft.structure.StructureTemplateManager;
+import net.minecraft.util.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
+import net.minecraft.world.gen.structure.Structure;
 import net.thesquire.backroomsmod.BackroomsMod;
 import net.thesquire.backroomsmod.config.ModConfig;
+import net.thesquire.backroomsmod.world.structure.ModStructureKeys;
 
 import java.util.Optional;
 
@@ -72,6 +79,58 @@ public class ModUtils {
         }
         BackroomsMod.LOGGER.warn("Unable to read property: {} with value: {} for blockstate: {}", key, properties.getString(key), root.toString());
         return state;
+    }
+
+    public static Optional<BlockPos> findStructure(ServerWorld world, BlockPos pos, RegistryKey<Structure> structureKey) {
+        Registry<Structure> registry = world.getRegistryManager().get(RegistryKeys.STRUCTURE);
+        Optional<RegistryEntry.Reference<Structure>> registryEntry = registry.getEntry(structureKey);
+        if(registryEntry.isEmpty()) return Optional.empty();
+
+        RegistryEntryList<Structure> registryEntryList = RegistryEntryList.of(registryEntry.get());
+        Pair<BlockPos, RegistryEntry<Structure>> pair = world.getChunkManager().getChunkGenerator()
+                .locateStructure(world, registryEntryList, pos, 100, false);
+        if (pair == null) return Optional.empty();
+
+        return Optional.of(pair.getFirst());
+    }
+
+    public static void placeStructure(ServerWorld world, String name, BlockPos pos, int angle) {
+        Optional<StructureTemplate> optional;
+        StructureTemplateManager structureTemplateManager = world.getStructureTemplateManager();
+        Identifier templateName = StringHelper.isEmpty(name) ? null : Identifier.tryParse(name);
+        if(templateName == null) {
+            BackroomsMod.LOGGER.warn("Invalid structure name: " + name);
+            return;
+        }
+
+        // get StructureTemplate from supplied name
+        try {
+            optional = structureTemplateManager.getTemplate(templateName);
+        }
+        catch (InvalidIdentifierException invalidIdentifierException) {
+            return;
+        }
+        if (optional.isEmpty()) {
+            return;
+        }
+        StructureTemplate template = optional.get();
+
+        // get BlockRotation from supplied angle
+        BlockRotation rotation;
+        switch (angle) {
+            case 90 -> rotation = BlockRotation.CLOCKWISE_90;
+            case 180 -> rotation = BlockRotation.CLOCKWISE_180;
+            case 270 -> rotation = BlockRotation.COUNTERCLOCKWISE_90;
+            default -> rotation = BlockRotation.NONE;
+        }
+
+        // for now, no mirroring and doesn't ignore entities
+        StructurePlacementData structurePlacementData = new StructurePlacementData()
+                .setRotation(rotation)
+                .setMirror(BlockMirror.NONE)
+                .setIgnoreEntities(false);
+
+        template.place(world, pos, pos, structurePlacementData, StructureBlockBlockEntity.createRandom(0L), Block.NOTIFY_LISTENERS);
     }
 
 }