+++ CHANGELOG.md
@@ -26,9 +26,11 @@
     - this tag does not save when closing the world, can only be written to
   - replace block uses same format as BlockStateTag described in Minecraft wiki
     - https://minecraft.fandom.com/wiki/Tutorials/Command_NBT_tags#Blocks
+  - can be configured to place an additional structure in the destination dimension
 - fixed ModSounds class related crash
 - re-did flat_no_caves.json noise settings to fix missing ceiling tiles
 - added level 1 portal structure
   - generates in level 0
   - contains portal placer block to place the portals
   - 8 block wide portal
+  - portal placer block places destination portal structure in level 1